requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asgi-csrf==0.11
asgiref==3.8.1
asn1crypto==1.5.1
astunparse==1.6.3
attrs==23.2.0
awscli==1.37.20
backoff==2.2.1
bcrypt==4.2.1
blinker==1.7.0
boto3==1.36.20
botocore==1.36.20
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
click-default-group==1.2.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==44.0.0
cvzone==1.6.1
cycler==0.12.1
dataclasses-json==0.6.7
datasette==0.65.1
db-utils==0.4.5
Deprecated==1.2.15
distlib==0.3.9
distro==1.9.0
Django==5.1.4
djangorestframework==3.15.2
dm-tree==0.1.8
dnspython==2.7.0
docutils==0.16
docx2txt==0.8
durationpy==0.9
easyocr==1.7.1
email_validator==2.2.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastapi-cli==0.0.6
fastembed==0.5.0
filelock==3.16.1
Flask==3.0.2
Flask-Cors==5.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
flexcache==0.3
flexparser==0.4
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.2
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.2
greenlet==3.0.3
groq==0.13.1
grpcio==1.68.1
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
hupper==1.12.1
hyperframe==6.0.1
idna==3.10
imageio==2.34.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
itsdangerous==2.1.2
janus==2.0.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.12.0
kiwisolver==1.4.5
kubernetes==31.0.0
labelImg==1.8.6
langchain==0.3.10
langchain-chroma==0.1.4
langchain-community==0.3.10
langchain-core==0.3.32
langchain-groq==0.2.3
langchain-huggingface==0.1.2
langchain-ollama==0.2.1
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.147
lazy_loader==0.3
libclang==16.0.6
loguru==0.7.3
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.8.3
mdurl==0.1.2
mergedeep==1.3.4
ml-dtypes==0.4.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgspec==0.19.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
namex==0.0.7
narwhals==1.16.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.3
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.57.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt-einsum==3.3.0
optree==0.13.1
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
posthog==3.7.4
proto-plus==1.26.0
protobuf==5.29.1
psutil==5.9.8
psycopg2==2.9.10
py-cpuinfo==9.0.0
py_rust_stemmers==0.1.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.11.1
pyodbc==5.2.0
pyOpenSSL==24.3.0
pyparsing==3.1.2
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.5.4
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ffmpeg==2.0.10
python-multipart==0.0.19
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rich-toolkit==0.12.0
rpds-py==0.22.3
rsa==4.7.2
s3transfer==0.11.2
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
SpeechRecognition==3.14.1
SQLAlchemy==2.0.29
sqlparse==0.5.3
starlette==0.41.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tesseract==0.1.3
tf_keras==2.18.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.2.12
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.2.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.0
twilio==9.0.4
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.1.29
uri==2.0.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
virtualenv==20.28.0
watchdog==6.0.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.1
win32_setctime==1.2.0
wrapt==1.14.1
WTForms==3.1.2
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.21.0
